Included device to device inter-node transfer in cxi_transfer_test. fixed some bugs related to device selection and cuda buffer mr registration TODO: check device to host RDMA and AMD transfer

+++ mooncake-transfer-engine/include/transport/transport.h
@@ -111,7 +111,7 @@ class Transport {
         std::string peer_nic_path;
         SliceStatus status;
         TransferTask *task;
-        std::vector<uint32_t> dest_rkeys;
+        std::vector<uint64_t> dest_rkeys;
         bool from_cache;
 
         union {
